+++ lib/calendlyr/resources/webhooks.rb
@@ -1,9 +1,14 @@
 module Calendlyr
   class WebhooksResource < Resource
     def list(organization:, scope:, **params)
+      next_page_caller = ->(page_token:) { list(organization: organization, scope: scope, **params, page_token: page_token) }
       organization = expand_uri(organization, "organizations")
       response = get_request("webhook_subscriptions", params: params.merge(organization: organization, scope: scope).compact)
-      Collection.from_response(response, type: Webhooks::Subscription, client: client)
+      Collection.from_response(response, type: Webhooks::Subscription, client: client, next_page_caller: next_page_caller)
+    end
+
+    def list_all(organization:, scope:, **params)
+      list(organization: organization, scope: scope, **params).auto_paginate.to_a
     end
 
     def create(url:, events:, organization:, scope:, **params)